+++ contracts/AssetTokenEL.sol
@@ -7,41 +7,43 @@ import "./IAssetToken.sol";
 import "./AssetTokenBase.sol";
 
 contract AssetTokenEL is IAssetTokenERC20, AssetTokenBase {
-
-    using SafeMath for uint;
+    using SafeMath for uint256;
 
     IERC20 private _el;
 
     /// @notice Emitted when an user claimed reward
-    event RewardClaimed(address account, uint reward);
+    event RewardClaimed(address account, uint256 reward);
 
     constructor(
         IERC20 el_,
         IEController eController_,
-        uint amount_,
-        uint price_,
-        uint rewardPerBlock_,
-        uint payment_,
-        uint latitude_,
-        uint longitude_,
-        uint assetPrice_,
-        uint interestRate_,
+        uint256 amount_,
+        uint256 price_,
+        uint256 rewardPerBlock_,
+        uint256 payment_,
+        uint256 latitude_,
+        uint256 longitude_,
+        uint256 assetPrice_,
+        uint256 interestRate_,
         string memory name_,
         string memory symbol_,
         uint8 decimals_
-    ) AssetTokenBase (
-        eController_,
-        amount_,
-        price_,
-        rewardPerBlock_,
-        payment_,
-        latitude_,
-        longitude_,
-        assetPrice_,
-        interestRate_,
-        name_,
-        symbol_,
-        decimals_) {
+    )
+        AssetTokenBase(
+            eController_,
+            amount_,
+            price_,
+            rewardPerBlock_,
+            payment_,
+            latitude_,
+            longitude_,
+            assetPrice_,
+            interestRate_,
+            name_,
+            symbol_,
+            decimals_
+        )
+    {
         _el = el_;
     }
 
@@ -54,10 +56,22 @@ contract AssetTokenEL is IAssetTokenERC20, AssetTokenBase {
      * - `amount` this contract should have more asset token than the amount.
      * - `amount` msg.sender should have more el than elAmount converted from the amount.
      */
-    function purchase(uint amount) external override whenNotPaused returns (bool) {
+    function purchase(uint256 amount)
+        external
+        override
+        whenNotPaused
+        returns (bool)
+    {
         _checkBalance(msg.sender, address(this), amount);
 
-        require(_el.transferFrom(msg.sender, address(this), eController.mulPrice(amount, price)), 'EL : transferFrom failed');
+        require(
+            _el.transferFrom(
+                msg.sender,
+                address(this),
+                eController.mulPrice(amount, price)
+            ),
+            "EL : transferFrom failed"
+        );
         transferFrom(address(this), msg.sender, amount);
 
         return true;
@@ -72,10 +86,18 @@ contract AssetTokenEL is IAssetTokenERC20, AssetTokenBase {
      * - `amount` msg.sender should have more asset token than the amount.
      * - `amount` this contract should have more el than elAmount converted from the amount.
      */
-    function refund(uint amount) external override whenNotPaused returns (bool) {
+    function refund(uint256 amount)
+        external
+        override
+        whenNotPaused
+        returns (bool)
+    {
         _checkBalance(address(this), msg.sender, amount);
 
-        require(_el.transfer(msg.sender, eController.mulPrice(amount, price)), 'EL : transfer failed');
+        require(
+            _el.transfer(msg.sender, eController.mulPrice(amount, price)),
+            "EL : transfer failed"
+        );
         transferFrom(msg.sender, address(this), amount);
 
         return true;
@@ -91,14 +113,23 @@ contract AssetTokenEL is IAssetTokenERC20, AssetTokenBase {
      * Requirements:
      * - `elPrice` cannot be the zero.
      */
-    function claimReward() external override onlyWhitelisted(msg.sender) whenNotPaused {
-        uint reward = getReward(msg.sender).mul(1e18).div(eController.getPrice());
+    function claimReward()
+        external
+        override
+        onlyWhitelisted(msg.sender)
+        whenNotPaused
+    {
+        uint256 reward =
+            getReward(msg.sender).mul(1e18).div(eController.getPrice());
 
-        require(reward < _el.balanceOf(address(this)), 'AssetToken: Insufficient seller balance.');
+        require(
+            reward < _el.balanceOf(address(this)),
+            "AssetToken: Insufficient seller balance."
+        );
         _el.transfer(msg.sender, reward);
         _clearReward(msg.sender);
 
-		emit RewardClaimed(msg.sender, reward);
+        emit RewardClaimed(msg.sender, reward);
     }
 
     /**
@@ -110,9 +141,19 @@ contract AssetTokenEL is IAssetTokenERC20, AssetTokenBase {
      * - `amount` buyer should have more asset token than the amount.
      * - `amount` seller should have more el than elAmount converted from the amount.
      */
-    function _checkBalance(address buyer, address seller, uint amount) internal view {
-        require(_el.balanceOf(buyer) > eController.mulPrice(amount, price), 'AssetToken: Insufficient buyer el balance.');
-        require(balanceOf(seller) > amount, 'AssetToken: Insufficient seller balance.');
+    function _checkBalance(
+        address buyer,
+        address seller,
+        uint256 amount
+    ) internal view {
+        require(
+            _el.balanceOf(buyer) > eController.mulPrice(amount, price),
+            "AssetToken: Insufficient buyer el balance."
+        );
+        require(
+            balanceOf(seller) > amount,
+            "AssetToken: Insufficient seller balance."
+        );
     }
 
     /**
@@ -121,6 +162,4 @@ contract AssetTokenEL is IAssetTokenERC20, AssetTokenBase {
     function withdrawToAdmin() public onlyAdmin(msg.sender) {
         _el.transfer(msg.sender, _el.balanceOf(address(this)));
     }
-
-
 }
